feat: give persistent mind image attachments a shared service owner
